Added support for BIV 1.5
BIV was updated with support to provide integrity for package.conf boots of IOS-XE
This means the PCR values need to be calculated over ALL of the packages.

In addition the KGV file had to be updated to include `biv_hash` of each sub package.

A couple other changes around testing to utilize `pathlib.Path` rather than building strings.
Hopefully the tests will work better on windows now (I haven't tested it)