Enhancement - Ability to toggle exiting on KGV mismatch
Fixes #7

Allows users to supply a boolean value for the new `ignore_kgv_match_failure` flag. This flag is used to determine
if a KGV mismatch should stop further validation and raise an exception, or simply log the mismatch error and continue
with subsequent validation steps.

Default behavior if a value for the flag is not supplied is the original mechanism of raising an exception.